+++ fixtures/application/sample1.tsx
@@ -1,0 +1,1 @@
+<tileset version="1.8" tiledversion="1.8.0" name="test" tilewidth="1" tileheight="1" tilecount="0" columns="0"/>