+++ src/macros.rs
@@ -1,6 +1,15 @@
-//! Macros.
+//! Macros for generating [FileFormat](crate::FileFormat) enum and associated methods.
 
-/// Generates `FileFormat` enum with methods.
+/// Generates the [FileFormat](crate::FileFormat) enum with methods for retrieving information.
+///
+/// Each file format includes the following properties:
+///
+/// - `format`: The variant name representing the file format.
+/// - `name`: The full name of the file format.
+/// - `short_name`: An abbreviated name for the file format (optional).
+/// - `media_type`: The common media type associated with the file format.
+/// - `extension`: The common file extension used for the file format.
+/// - `kind`: The type or category of the file format.
 macro_rules! formats {
     {
         $(
@@ -122,7 +131,13 @@ macro_rules! formats {
     };
 }
 
-/// Generates `FileFormat::from_signature` function.
+/// Generates the [FileFormat::from_signature](crate::FileFormat::from_signature) function.
+///
+/// Each signature group includes the following properties:
+///
+/// - `format`: The variant name representing the file format.
+/// - `value`: The signature value associated with the format (can be repeated).
+/// - `offset`: The offset to start matching the signature value (defaults to 0 if not specified).
 macro_rules! signatures {
     {
         $(