+++ tests/document.rs
@@ -224,3 +224,147 @@ fn test_rich_text_format() {
     let fmt = FileFormat::from_file("fixtures/document/sample.rtf").unwrap();
     assert_eq!(fmt, FileFormat::RichTextFormat);
 }
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_starcalc() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sdc").unwrap();
+    assert_eq!(fmt, FileFormat::Starcalc);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_starchart() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sds").unwrap();
+    assert_eq!(fmt, FileFormat::Starchart);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_stardraw() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sda").unwrap();
+    assert_eq!(fmt, FileFormat::Stardraw);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_starimpress() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sdd").unwrap();
+    assert_eq!(fmt, FileFormat::Starimpress);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_starmath() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.smf").unwrap();
+    assert_eq!(fmt, FileFormat::Starmath);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_starwriter() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sdw").unwrap();
+    assert_eq!(fmt, FileFormat::Starwriter);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_calc() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sxc").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlCalc);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_calc_template() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.stc").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlCalcTemplate);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_draw() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sxd").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlDraw);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_draw_template() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.std").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlDrawTemplate);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_impress() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sxi").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlImpress);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_impress_template() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sti").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlImpressTemplate);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_math() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sxm").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlMath);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_writer() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sxw").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlWriter);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_writer_global() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.sgw").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlWriterGlobal);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_sun_xml_writer_template() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.stw").unwrap();
+    assert_eq!(fmt, FileFormat::SunXmlWriterTemplate);
+}
+
+#[test]
+fn test_wordperfect_document_1() {
+    let fmt = FileFormat::from_file("fixtures/document/sample1.wpd").unwrap();
+    assert_eq!(fmt, FileFormat::WordperfectDocument);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_wordperfect_document_2() {
+    let fmt = FileFormat::from_file("fixtures/document/sample2.wpd").unwrap();
+    assert_eq!(fmt, FileFormat::WordperfectDocument);
+}
+
+#[test]
+fn test_wordperfect_graphics_1() {
+    let fmt = FileFormat::from_file("fixtures/document/sample1.wpg").unwrap();
+    assert_eq!(fmt, FileFormat::WordperfectGraphics);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
+fn test_wordperfect_graphics_2() {
+    let fmt = FileFormat::from_file("fixtures/document/sample2.wpg").unwrap();
+    assert_eq!(fmt, FileFormat::WordperfectGraphics);
+}
+
+#[test]
+fn test_wordperfect_macro() {
+    let fmt = FileFormat::from_file("fixtures/document/sample.wpm").unwrap();
+    assert_eq!(fmt, FileFormat::WordperfectMacro);
+}